refactor: refactor ChatInput
- Extract onSubmit to props.
- Focus on route param change instread of prop change.

After the refactoring the two ChatInput components are more similiar, and may be someday I can extract a common one.
